Ideal Modern Syntax
This is what the ideal syntax would look like, only if it meets the compatibility goals that project needs. I'm not sure what those are yet, so I'm adding these just to see how it looks to start with.

Here's the issue I opened that's associated with this, learning more about what the compatibility goals are.

https://github.com/101arrowz/compression-streams-polyfill/issues/1

I tried this setup out in my own app Dovetail, and you can enable the polyfill by simply side-effect importing from the polyfill module, like this:

```ts
import "compression-streams-polyfill/ponyfill";
```

While using Import Maps in the browser, you can use that import as-is, with it simply mapping to wherever you would like to load the polyfill from (Your own `node_modules` folder, or a CDN).